+++ src/main/kotlin/net/dv/tax/dto/sales/CarInsuranceDto.kt
@@ -2,7 +2,7 @@ package net.dv.tax.dto.sales
 
 data class CarInsuranceDto(
 
-    val year: Int,
+    val year: Long,
 
     val month: Int
